Move utils folder into src
jasina, tjha
Also, fix corresponding includes
