--- src/example.cpp
@@ -1,16 +1,0 @@
-// example.cpp
-// 
-// Example "Hello World" program to test Travis CI workflow
-//
-// History:
-//    - 12 September 2019: Tejas Jha implemented "Hello World"
-#include<iostream>
-#include"../utils/vector.hpp"
-
-
-int main()
-   {
-   std::cout << "Hello World!" << std::endl;
-   return 0;
-   }
-
